mnt: Fix mntflush "rpc tags" error
This might be the third time we tried to fix this.  In this case, if a
process has a syscall in progress that is blocked in mntio(), then when
that process dies, its syscall will get aborted.  When this loop retried,
it would block and immediately fail again (rendez aborts when DYING).

Since this loop didn't allow aborts to happen (abort == retry, here), we
needed to special case when the process is actually dying and *really*
abort.

Ugh.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>