devroot: Use dirtab's length instead of rootdata's size
There were two mechanisms tracking the size of an object in #root - the
dirtab's length and the rootdata's size fields.  We didn't always keep them
in sync, and we really only need one.

I also removed the sizep field - it wasn't clear that it was used for
anything other than the rootinit() attempt to keep the length and size in
sync.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>