WIP-9ns
-------- probably don't want this.  get a real RAM FS instead of using a
dirtab.  rename and #root wstat is all messed up too.

9ns: Increase KNAMELEN from 28

This allows 9ns files to have names larger than 28 bytes, especially
useful in #root.  Arguably we wouldn't need this if we weren't using a
lousy dirtab.

160 + 32 (the rest of the dirtab) is 64 byte aligned.  I wanted 128
bytes of space for the KNAME, and figured we should use the rest of the
space.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>