vmrunkernel: get the page table setup right
I screwed this up so many times, maybe this time it's right.
Tested on 3 different memory sizes with weird alignments.

I've added a new check, checkmemaligned, which ensures
that memstart and memsize are 2m aligned.

This check needs to be separate from the memory
function as some environments may not call
memory() but still need to check alignment.

Change-Id: Ie7b8e324683f0415bd8d5d078fe3ff29a04a515e
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>
Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>